pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "copilotkit>=0.1.79",
    "fastapi>=0.135.1",
    "langchain>=0.3.0",
    "langchain-core>=1.2.19",
    "langchain-openai>=0.3.0",
    "langgraph>=0.4.0",
    "langgraph-checkpoint>=4.0.1",
    "ag-ui-langgraph>=0.0.20",
    "python-dotenv>=1.2.2",
    "starlette>=0.52.1",
    "uvicorn>=0.42.0",
]
